changed cardUI to event monobehaviour
