+++ Assets/Scripts/Cards/CardEvent.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: a8470e08e90f49f48a9ac7b0642daef8
+timeCreated: 1582273709
