+++ Assets/Scripts/Cards/Render.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections;
 using System.Collections.Generic;
 using _Editor;
@@ -5,16 +6,18 @@ using UnityEngine;
 using TMPro;
 
 using Managers;
+using UI;
 
 namespace Cards
 {
     public class Render : MonoBehaviour
     {
+        public bool visible = true;
+        
+        public float moveSpeed = 0.1f;
+
         public void Start()
         {
-            //cardui.setCallbacks();
-            int sortOrder;
-
             MetaData meta = GetComponent<MetaData>();
             
             // Set strategy color
@@ -31,7 +34,7 @@ namespace Cards
             var strRenderer = transform.Find("StrategySprite").GetComponent<SpriteRenderer>();
             strRenderer.sprite = Resources.Load<Sprite>(VfxManager.StrategySpritePaths[meta.strategy]);
 
-            sortOrder = Game.Ctx.VfxOperator.GetSortOrder();
+            var sortOrder = Game.Ctx.VfxOperator.GetSortOrder();
             var canvas = GetComponent<Canvas>();
             canvas.sortingLayerName = "Card";
             canvas.sortingOrder = sortOrder;
@@ -48,5 +51,46 @@ namespace Cards
             // Set rules text
             transform.Find("CardText").GetComponent<TextMeshProUGUI>().text = GetComponent<MetaData>().description;
         }
+
+        IEnumerator MoveCard(Vector3 dest, float delay = 0)
+        {
+            Vector3 init = new Vector3(transform.position.x, transform.position.y);
+            float t = 0f;
+            yield return new WaitForSeconds(delay);
+            while (t < moveSpeed) {
+                float i = t / moveSpeed;
+                transform.SetPositionAndRotation(i * dest + (1f - i) * init,
+                    transform.rotation);
+                t += Time.deltaTime;
+                yield return null;
+            }
+        }
+    
+        public void Hide()
+        {
+            visible = false;
+            TextMeshProUGUI[] tmPros = GetComponentsInChildren<TextMeshProUGUI>();
+            foreach (var r in tmPros) {
+                r.enabled = false;
+            }
+            SpriteRenderer[] spRenderers = GetComponentsInChildren<SpriteRenderer>();
+            foreach (var r in spRenderers) {
+                r.enabled = false;
+            }
+        }
+    
+        public void Show()
+        {
+            visible = true;
+            TextMeshProUGUI[] tmPros = GetComponentsInChildren<TextMeshProUGUI>();
+            foreach (var r in tmPros) {
+                r.enabled = true;
+            }
+            SpriteRenderer[] spRenderers = GetComponentsInChildren<SpriteRenderer>();
+            foreach (var r in spRenderers) {
+                r.enabled = true;
+            }
+        }
     }
+    
 }
