--- Assets/Scripts/UI/CardUI.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c9a6038e7d8a7bb4dbc91d989b96a537
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
